refactor(integrations): pre-1.0 API cleanup — unleak css-in-js internals, name TailwindIntegrationOptions.roles
